requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
mysql-connector-python>=9.1.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
